update: move useMenuPosition hook to a different folder
